refactor(client): separate auth from client implementation
