+++ src/client/errors.rs
