+++ src/client/tests/auth_error.json
