+++ src/provisioning/models.rs
@@ -7,7 +7,7 @@ use reqwest::Response;
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    auth::{DracoonClient, DracoonErrorResponse},
+    client::{DracoonClient, DracoonErrorResponse},
     user::UserAuthData,
     utils::{parse_body, FromResponse},
     DracoonClientError, KeyValueEntry, RangedItems,
