+++ src/roles/models.rs
@@ -5,7 +5,7 @@ use reqwest::Response;
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    auth::{DracoonClient, DracoonErrorResponse},
+    client::{DracoonClient, DracoonErrorResponse},
     models::RangedItems,
     nodes::UserInfo,
     utils::{parse_body, FromResponse},
