+++ src/shares/models/download.rs
@@ -6,7 +6,7 @@ use reqwest::Response;
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    auth::DracoonErrorResponse,
+    client::DracoonErrorResponse,
     models::{FilterOperator, FilterQuery, ObjectExpiration, RangedItems, SortOrder, SortQuery},
     nodes::models::{NodeType, UserInfo},
     utils::{parse_body, FromResponse},
