+++ src/shares/models/upload.rs
@@ -5,7 +5,7 @@ use reqwest::Response;
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    auth::DracoonErrorResponse,
+    client::DracoonErrorResponse,
     models::{FilterOperator, FilterQuery, ObjectExpiration, RangedItems, SortOrder, SortQuery},
     nodes::models::UserInfo,
     utils::{parse_body, FromResponse},
